changes in validation in placa

+++ src/routes/deletePlaca.js
@@ -10,11 +10,13 @@ const { Vehiculo } = db.models;
 router.delete('/:placa', async (req, res) => {
 
     try {
-        
+
         const { placa } = req.params;
+        if (placa.length < 5 || placa.length > 6) return res.status(404).json({ msg: 'Placa incompleta' });
+
         let consulta = await Vehiculo.findOne({ where: { placa: placa } });
         if (!consulta) return res.status(404).json({ msg: 'No existe esta placa en BD.' });
-        
+
         await Vehiculo.destroy({ where: { placa: placa } });
         return res.json({ msg: 'Placa eliminada.' })
